+++ .luacov
@@ -30,7 +30,7 @@ return {
   -- all will be included if nothing is listed
   -- (exclude overrules include, do not include
   -- the .lua extension, path separator is always '/')
-  ["include"] = {  },
+  ["include"] = { "./src"  },
 
   -- Patterns for files to exclude when reporting
   -- all will be included if nothing is listed
